Add getHoppingMatrix in BandStructure.jl.
The getHoppingMatrix function is added to the BandStructure.jl module.
This function calculates the hopping matrix from the unit cell at the
origin to one of the neighboring unit cells labeled by an integer.
Valid labels for a unit cell are in the range 0:2 * dim. The label 0
designates the unit cell at the origin, the labels 1:dim designate the
unit cells given by the primitive vectors, bravaisvec[:,1:dim] and the
labels in the range dim + 1:2 * dim desiginate the antiparallel
primitive vectors -bravaisvec[:, 1:dim].

Again only hopping matrices from the unit cell at the origin to another
neighboring unit cell are considered because these are the only matrices
that emerge in the calculation of the bloch matrices in a tight binding
model.