Fix two errors in CrystalLattice.jl
The first of these errors is the order of the arguments in the iterator
over all unit cells neighboring the unit cell at the origin. The
function getPrimitiveNearestNeighbors ordered the bravais vectors and
primitive vectors in the opposite manner from which they were parsed in
the getNearestNeighbors method. This was changed so that they coincide
and so that the base vectors remain the value which the iterator
changes fastest.

The second of these errors was more technical and had to do with calling
the method isapprox(d - min, 0) opposed to isapprox(d, min). Calling
isapprox the first way will almost always result in false if no
absolute tolerance value is defined (for an explanation as to why please
see the julia documentation) for this reason the call was changed to the
second (isapprox(d, min)).