A basic electron lattice module was added.
This module serves the purpose of iteratively adding structure to the
underlying bravais lattice. The bravais lattice is the most basic
structure consisting of simply the bravais lattice vectors, the crystal
structure consists of a bravais lattice with a basis, and the electron
lattice is a population of electrons with a given number of states at
each site.