Add getHoppingMatrixStructure to ElectronLattice.jl.
This function utilizes an array designating the nearest neighbors of
each element in the unit cell located at the origin in order to derive
the structure of the hopping matrix between this unit cell and a
neighboring one, given by the 'cell' variable passed to the function (0
is the unit cell at the origin, 1 is the unit cell given by the first
primitive vectors etc; the unit cells given by the negative i'th
primitive vector is labeled with n + i where n is the total number of
primitive vectors describing the system.). It should be emphasized that
only the structure of the hopping matrix is derived, with each unique
term in the hamiltonian identified by a prime number and each element of
the hopping matrix returned by this function consisting of a
multiplication of the prime numbers corresponding to the terms required
at the given lattice position.