Add module and test for the new struct FieldCorrelator.
The new module MeanField.jl is tasked with instantiating the appropriate
mean field correlation matrix and has a number of methods that attempt
to abstract this process out to easily allow the instantiation of these
objects regardless of the interactions present in the given model. The
module brakes the instantiation process into three steps. Firstly, it
instantiates a Matrix of the appropriate size as determined by the type
of interactions present in the system. Then, it assigns populates this
matrix with prime numbers which label appropriate values to be used
during instantiation. Finally it populates the matrix of prime numbers
with the appropriate values corresponding to the prime labels.

This decoupling of matrix elements from values allows the permutation
operations to be carried out on the prime:value dictionary to change the
instantiation method (ex: ferromagnetic, antiferromagnetic).