Add ChemicalPotential.jl and test module.
ChemicalPotential.jl utilizes an implementation of the bisection
algorithm provided by Roots.jl in order to solve for the chemical
potential of a given crystal provided a given filling factor. The test
module uses the functions defined in ChemicalPotential.jl to validate
they return a chemical potential of 0.0 for single layer non-interacting
graphene at half filling.