Change configs to .toml, modify constructors accordingly.
This change has many benefits. Firstly, the .toml file has a
specification ensuring that the way in which various values are
configured in this package is not completely unique to this package.
Secondly, the .toml specification can handle arrays of floats which the
.ini file could only do with hacked together methods. Thirdly, the TOML
package shipped with julia is a lot nicer to work with then  the
Configs.jl package (Everything works as expected upon calling the
TOML.parse_file functions).

The other change that is mixed in here is that size and population
related fields of the ElectronLattice.jl and CrystalLattice.jl modules
were removed. There is no use for them in the current scope of this
package.