Add functions for generating the bravais lattice in BravaisLattice.jl
The getLatticePoints function was added in BravaisLattice.jl to which
the parameter 'order' has to be passed. This parameter determines how
far out the bravais lattice is generated from the origin. If 1 is passed
then simply the bravais lattice point at the origin and all points
corresponding to a single step along a direction parallel to one of the
bravais vectors are generated. If 2 is passed then all the points are
generated that would be generated in 1 as well as all points one step
away from the boundary points by a simple bravais vector.

A simple function is added to the CrystalInitTest.jl class that
visualizes the generated points for a triangle lattice, to validate that
the points are generated correctly.