Add getPoints and getNearestNeighborsTest to CrystalLattice.jl
In this commit we add the requisite functions to CrystalLattice.jl in
order to utilize the NearestNeighbors.jl package added previously. These
functions are the getPoints and getNearestNeighborsTest functions.
getNearestNeighborsTest was named so as not to interfere with the
already existing getNearestNeighbors function that was created for
non-interacting graphene.

The getPoints function returns the bravaislatticepoints and crystalsites
up to a given distance away from the origin as determined by the 'order'
parameter (for an explanation on how the 'order' determines how far out
from the origin these points are generated, please see the previous
commit).

The getNearestNeighborsTest function takes a set of points and returns
the k-nearest-neigbors of each of the atoms in the primitive unit cell
located at the origin.

A sanity check function is added to CrystalInitTest.jl to ensure that I
understand the output of the getNearestNeighborsTest function. This
function is a sanity check and not a true test, for this reason it's
call in the CrystalInitTest.jl module is commented out.