Only allow specification of the coloumb interaction with a vector.
The coulomb interaction can no longer be defined as a floating point
number in the .toml config file. If the coulomb interaction is only
specified for its nearest neighbors then this is specified as a single
element vector i.e. a 'bracketed' float. This is done to simplify the
code necessary to parse the appropriate information from this field.