Make latticepoints a field of the Bravais struct in BravaisLattice.jl
The getLatticePoints function was renamed to initlatticepoints and its
parameters were changed so that the function could be utilized in the
constructor of the Bravais struct to initialize the lattice points as a
field of the struct itself. To this end, we furthermore had to move the
order parameter to the bravais config file.

The reason this change was made, is because of the centrality of lattice
points to the tight binding model in this program. The arrangement of
these points in the latticepoints matrix defines the order of iteration
over the lattice and thus defines the order in which all other relevant
elements of the hamiltonian should be initialized.