Generalize the functions present in Embeddings.jl
The functions in Embeddings.jl assumed, in embedding an iterator, whose
elements were indices of arbritrary length ((i,j,k,l,...) or for example
(i,j,k,l)), that the length of the iterators elements was even and that
the range of each element in the first half ((i,j) in (1:3, 1:2)) was the
same as each one in the second ((k,l) in (1:3, 1:2)). This commit removes
this assumption and allows the length of the elements of the iterator to
be either even or odd and makes no assumptions about the ranges of the
individual indices.

In line with these changes the EmbeddingsTest.jl module had to be
changed, so that the tests still passed.