Overhaul MeanField.jl to be in agreement with latticepoints.
The FieldCorrelator constructor was altered so that the shape of the
correlator field would always be in agreement with the shape of the
latticepoints field present in the Crystal struct. I do not feel like
explaining why I did this at the moment, but it has to do with being
able to generate the correlator matrix without having to take into
account the specific model.

The module MeanFieldTest.jl had to be changed accordingly.