Compute and label nearest neighbors with their respective order.
getNearestNeighborsTest was modified to no longer return the matrix
describing the euclidean distance between each point specified in the
search and the points in the search tree, but instead to return a matrix
of labels which describe what order of neighbors each of the points are
(i.e. first nearest neighbors, second nearest neighbors etc.). This
change was made because it is ultimately this information which is
relevant to determining if a given entry in the hopping matrix should
have an extra term added to it which is not present in the
non-interacting model.