Remove error check in BandStructureBeta.jl
This check assumes that a given element of the correlator will not be
initialized to zero if the state associated with it is associated with a
coulomb interaction term in the hamiltonian. This is however a false
assumption for the hubbard model where all elements of the correlation
matrix associated with states localized outside of the initial unit cell
are initialized to zero and the coulomb interaction may be specified as
an empty list or a list with a zero strength coulomb interaction
specified for nearest neighbors.