Perform the chemical potential update with appropriate energies.
Before this commit the update of correlator and the calculation of the
eigenenergies corresponding to the hamiltonian was performed in
parallel. Following this step the chemical potential was updated. This
order of operations is flawed because this means that for each
correlator update only the chemical potential belonging to the previous
iteration is present, when it should be the chemical potential of the
current iteration/corrrelator which is used to update it.