Specify types of energies parameter in ChemicalPotential.jl
In ChemicalPotential.jl the type of the 'energies' parameter was
specified to be an array. This was done because the information to be
passed through energies, that being the energy at each point in
reciprocal space for all k-points could either be specified as an
Array{Float64} or as a Vector{Vector{Float64}} depending on how the
energies are calculated. However, the length function called in this
module works differently depending on which type expresses this
information. We therefore require the 'energy' parameter to be of a
specific type in this case an array.