Remove constructHoppingMatrixTest from tests.
This test was intended to test an older version of getHoppingMatrixTest
and not the newer one. Furthermore, there doesn't seem to be to much
sense in writing a newer version now because we know the function works
and test are supposed to test behaviour and not simply whether or not
the function works correctly or not.